fix implement repository

+++ models/category.go
@@ -1,7 +1,7 @@
 package models
 
 type Category struct {
-	CategoryID   int64  `json:"category_id"`
+	CategoryID   int    `json:"category_id"`
 	CategoryName string `json:"category_name"`
 	Details      string `json:"details"`
 }
